refactor: animation hooks refactoring
