+++ ts/components/ui/utils/hooks/useSpringPressProgressValue.ts
@@ -1,0 +1,28 @@
+import * as React from "react";
+import Animated, {
+  useDerivedValue,
+  useSharedValue,
+  withSpring
+} from "react-native-reanimated";
+import { IOSpringValues } from "../../../core/variables/IOAnimations";
+
+export const useSpringPressProgressValue = (
+  springValue = IOSpringValues.button
+) => {
+  const isPressed: Animated.SharedValue<number> = useSharedValue(0);
+
+  const progress = useDerivedValue(() =>
+    withSpring(isPressed.value, springValue)
+  );
+
+  const onPressIn = React.useCallback(() => {
+    // eslint-disable-next-line functional/immutable-data
+    isPressed.value = 1;
+  }, [isPressed]);
+  const onPressOut = React.useCallback(() => {
+    // eslint-disable-next-line functional/immutable-data
+    isPressed.value = 0;
+  }, [isPressed]);
+
+  return { onPressIn, onPressOut, progress };
+};
